@@ -30,15 +30,15 @@
     "iso-639-1": "^1.2.1",
     "languagedetect": "^1.2.0",
     "location-history": "^1.0.0",
-    "material-ui": "^0.17.0",
+    "material-ui": "^0.20.2",
     "mkdirp": "^0.5.1",
     "music-metadata": "^3.1.0",
     "network-address": "^1.1.0",
     "parse-torrent": "^6.0.1",
     "prettier-bytes": "^1.0.1",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
-    "react-tap-event-plugin": "^2.0.1",
+    "prop-types": "^15.6.2",
+    "react": "^16.5.2",
+    "react-dom": "^16.5.2",
     "rimraf": "^2.5.2",
     "run-parallel": "^1.1.6",
     "semver": "^5.6.0",
