@@ -147,7 +147,7 @@ module.exports = class TorrentList extends React.Component {
         }
       }
       return (
-        <div style={styles.wrapper}>
+        <div key='progress-bar' style={styles.wrapper}>
           <LinearProgress style={styles.progress} mode='determinate' value={progress} />
         </div>
       )
@@ -200,7 +200,7 @@ module.exports = class TorrentList extends React.Component {
       const minutesStr = (hours || minutes) ? minutes + 'm' : ''
       const secondsStr = seconds + 's'
 
-      return (<span>{hoursStr} {minutesStr} {secondsStr} remaining</span>)
+      return (<span key='eta'>{hoursStr} {minutesStr} {secondsStr} remaining</span>)
     }
 
     function renderTorrentStatus () {
@@ -370,16 +370,16 @@ module.exports = class TorrentList extends React.Component {
 
     return (
       <div key='radial-progress' className={'radial-progress ' + cssClass}>
-        <div key='circle' className='circle'>
-          <div key='mask-full' className='mask full' style={transformFill}>
-            <div key='fill' className='fill' style={transformFill} />
+        <div className='circle'>
+          <div className='mask full' style={transformFill}>
+            <div className='fill' style={transformFill} />
           </div>
-          <div key='mask-half' className='mask half'>
-            <div key='fill' className='fill' style={transformFill} />
-            <div key='fill-fix' className='fill fix' style={transformFix} />
+          <div className='mask half'>
+            <div className='fill' style={transformFill} />
+            <div className='fill fix' style={transformFix} />
           </div>
         </div>
-        <div key='inset' className='inset' />
+        <div className='inset' />
       </div>
     )
   }
@@ -393,7 +393,7 @@ function getErrorMessage (torrentSummary) {
   const err = torrentSummary.error
   if (err === 'path-missing') {
     return (
-      <span>
+      <span key='path-missing'>
         Path missing.<br />
         Fix and restart the app, or delete the torrent.
       </span>
